+++ agent/agent-plugins/logging-log4j2/src/main/java/org/bithon/agent/plugin/log4j2/Log4j2Plugin.java
@@ -17,8 +17,8 @@
 package org.bithon.agent.plugin.log4j2;
 
 import org.bithon.agent.core.aop.descriptor.InterceptorDescriptor;
-import org.bithon.agent.core.aop.descriptor.MatcherUtils;
 import org.bithon.agent.core.aop.descriptor.MethodPointCutDescriptorBuilder;
+import org.bithon.agent.core.aop.matcher.Matchers;
 import org.bithon.agent.core.plugin.IPlugin;
 
 import java.util.Arrays;
@@ -42,11 +42,11 @@ public class Log4j2Plugin implements IPlugin {
             forClass("org.apache.logging.log4j.core.Logger")
                 .methods(
                     MethodPointCutDescriptorBuilder.build()
-                                                   .onMethod(MatcherUtils.withName("logMessage")
-                                                                         .and(MatcherUtils.takesArgument(1,
-                                                                                                         "org.apache.logging.log4j.Level"))
-                                                                         .and(MatcherUtils.takesArgument(4,
-                                                                                                         "java.lang.Throwable")))
+                                                   .onMethod(Matchers.withName("logMessage")
+                                                                     .and(Matchers.takesArgument(1,
+                                                                                                 "org.apache.logging.log4j.Level"))
+                                                                     .and(Matchers.takesArgument(4,
+                                                                                                 "java.lang.Throwable")))
                                                    .to("org.bithon.agent.plugin.log4j2.interceptor.LoggerLogMessage")
                 ),
 
@@ -56,17 +56,17 @@ public class Log4j2Plugin implements IPlugin {
                      * {@link org.apache.logging.log4j.core.pattern.PatternParser#PatternParser(org.apache.logging.log4j.core.config.Configuration, String, Class, Class)}
                      */
                     MethodPointCutDescriptorBuilder.build()
-                                                   .onConstructor(MatcherUtils.takesArguments(4))
+                                                   .onConstructor(Matchers.takesArguments(4))
                                                    .to("org.bithon.agent.plugin.log4j2.interceptor.PatternParserCtor"),
 
                     /**
                      * {@link org.apache.logging.log4j.core.pattern.PatternParser#parse(String, List, List, boolean, boolean, boolean)}
                      */
                     MethodPointCutDescriptorBuilder.build()
-                                                   .onMethod(MatcherUtils.withName("parse")
-                                                                         .and(MatcherUtils.takesArgument(0,
-                                                                                                         "java.lang.String"))
-                                                                         .and(MatcherUtils.takesArguments(6)))
+                                                   .onMethod(Matchers.withName("parse")
+                                                                     .and(Matchers.takesArgument(0,
+                                                                                                 "java.lang.String"))
+                                                                     .and(Matchers.takesArguments(6)))
                                                    .to("org.bithon.agent.plugin.log4j2.interceptor.PatternParserParse")
                 )
         );
