+++ agent/agent-plugins/spring-webflux/src/main/java/org/bithon/agent/plugin/spring/webflux/SpringWebFluxPlugin.java
@@ -17,8 +17,8 @@
 package org.bithon.agent.plugin.spring.webflux;
 
 import org.bithon.agent.core.aop.descriptor.InterceptorDescriptor;
-import org.bithon.agent.core.aop.descriptor.MatcherUtils;
 import org.bithon.agent.core.aop.descriptor.MethodPointCutDescriptorBuilder;
+import org.bithon.agent.core.aop.matcher.Matchers;
 import org.bithon.agent.core.plugin.IPlugin;
 
 import java.util.Arrays;
@@ -59,8 +59,8 @@ public class SpringWebFluxPlugin implements IPlugin {
             forClass("reactor.netty.http.server.HttpServerOperations")
                 .methods(
                     MethodPointCutDescriptorBuilder.build()
-                                                   .onConstructor(MatcherUtils.takesArguments(1)
-                                                                              .or(MatcherUtils.takesArguments(10)))
+                                                   .onConstructor(Matchers.takesArguments(1)
+                                                                          .or(Matchers.takesArguments(10)))
                                                    .to("org.bithon.agent.plugin.spring.webflux.interceptor.HttpServerOperations$Ctor")
                 ),
